requirements.txt
flask
flask-cors
flask-sqlalchemy
psycopg2-binary
pyaudio
python-dotenv
redis
requests
